Make ChildBudgetBadge separate component
